🚚 refactor templates + configs structure
