+++ external/anomaly/configs/base/configuration_enums.py
