+++ external/anomaly/configs/padim/__init__.py
