+++ external/anomaly/configs/padim/compression_config.json
