+++ external/anomaly/configs/padim/configuration.yaml
