+++ external/anomaly/configs/padim/pot_optimization_config.json
