+++ external/anomaly/configs/stfpm/compression_config.json
