+++ external/anomaly/configs/stfpm/configuration.yaml
