+++ external/anomaly/templates/segmentation/stfpm/template.yaml
@@ -1,14 +1,14 @@
 # Description.
-model_template_id: ote_anomaly_classification_stfpm
+model_template_id: ote_anomaly_segmentation_stfpm
 name: STFPM
-task_type: ANOMALY_CLASSIFICATION
+task_type: ANOMALY_SEGMENTATION
 task_family: VISION
 instantiation: "CLASS"
 summary: Use this model when the position of the objects in the image frame might differ between images.
 application: ~
 
 # Algo backend.
-framework: OTEAnomalyClassification v0.1.0
+framework: OTEAnomalyClassification v0.1.0 # TODO: update after the name has been changed on the platform side
 
 # Task implementations.
 entrypoints:
@@ -18,7 +18,7 @@ entrypoints:
 
 # Hyper Parameters
 hyper_parameters:
-  base_path: ../../../templates/common/stfpm/configuration.yaml
+  base_path: ../../../configs/stfpm/configuration.yaml
   parameter_overrides:
     learning_parameters:
       train_batch_size:
