+++ external/anomaly/tools/sample.py
@@ -344,7 +344,7 @@ def parse_args() -> Namespace:
     )
     parser.add_argument(
         "--model_template_path",
-        default="./configs/templates/classification/padim/template.yaml",
+        default="./templates/classification/padim/template.yaml",
     )
     parser.add_argument("--dataset_path", default="./datasets/MVTec")
     parser.add_argument("--category", default="bottle")
